feat: Moved frontend from Angular 4 to Angular 6 ! (#50)
This pull request introduce a new version of the FrontEnd that use Angular 6 and angular-cli.

This closes #37. The dist bundle is now significantely smaller than before !

All the modifications are invisible to the end user.
Migrated the calls to the Http module to the new and simpler HttpClient.
Adopted new import syntax from `rxjs`.
Upgraded templates to newest version of VMWare clarity.